Moving Unit tests under io.personium.core.* package
