+++ src/test/java/io/personium/core/model/impl/es/doc/PropertyAliasTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.personium.test.unit.core.model.impl.es.doc;
+package io.personium.core.model.impl.es.doc;
 
 import static org.junit.Assert.assertEquals;
